server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "authlib>=1.6.5",
    "bcrypt==3.2.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "coloredlogs==15.0.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "cython==3.2.2",
    "easydict==1.13",
    "fastapi>=0.123.4",
    "flatbuffers==25.9.23",
    "fonttools==4.61.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "humanfriendly==10.0",
    "idna>=3.11 ; python_full_version >= '3'",
    "imageio==2.37.2",
    "insightface",
    "itsdangerous>=2.2.0",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "langchain>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "lazy-loader==0.4",
    "matplotlib==3.10.7",
    "ml-dtypes==0.5.4",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy>=0.38.0",
    "onnx==1.20.0",
    "onnxruntime==1.23.2",
    "openai-whisper>=20250625",
    "opencv-python==4.9.0.80",
    "packaging==25.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow==12.0.0",
    "prettytable==3.17.0",
    "protobuf==6.33.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "simsimd==6.5.3",
    "six==1.17.0",
    "sounddevice>=0.5.3",
    "sqlalchemy>=2.0.44",
    "stringzilla==4.4.0",
    "sympy==1.14.0",
    "tifffile==2025.5.10",
    "tqdm==4.67.1",
    "transformers>=4.57.3",
    "urllib3==2.5.0",
    "uvicorn>=0.38.0",
    "wcwidth==0.2.14",
    "webrtcvad-wheels>=2.0.14",
]
